Minor stylistic changes

+++ src/arguments.cpp
@@ -19,6 +19,7 @@
 #include <string.h>
 #include "arguments.h"
 
+
 // Predefined value that denotes successful operation
 const Error Error::OK(NULL);
 
@@ -40,7 +41,7 @@ const Error Error::OK(NULL);
 //     traces[=N]    - dump top N call traces
 //     flat[=N]      - dump top N methods (aka flat profile)
 //     interval=N    - sampling interval in ns (default: 1'000'000, i.e. 1 ms)
-//     jstackdepth=N - java stack depth (default: MAX_STACK_FRAMES, will be ignored if the option is greater than default)
+//     jstackdepth=N - maximum Java stack depth (default: MAX_STACK_FRAMES)
 //     framebuf=N    - size of the buffer for stack frames (default: 1'000'000)
 //     threads       - profile different threads separately
 //     simple        - simple class names instead of FQN
@@ -96,7 +97,7 @@ Error Arguments::parse(const char* args) {
                 return Error("interval must be > 0");
             }
         } else if (strcmp(arg, "jstackdepth") == 0) {
-            if (value == NULL || (_jstackdepth = atol(value)) <= 0) {
+            if (value == NULL || (_jstackdepth = atoi(value)) <= 0) {
                 return Error("jstackdepth must be > 0");
             }
         } else if (strcmp(arg, "framebuf") == 0) {
